Update high resoultion media query
This commit adds a the dppx variant of min-resolution to add an easier
to use value for the media query. `dppx` is not supported by all
browsers and needs the `dpi` fallblack.
`dppx` is more acurate than `dpi` since dots per inch refers to a
defined size of pixels wher one inch is 96px what so ever and does not
take physical inches into account.
Autoprefixer will include `-webkit-min-device-pixel-ratio` which is
needed for Safari and WebKit devices.

Reference #1474.